+++ esp-hal-common/src/soc/esp32s3/radio_clocks.rs
@@ -1,11 +1,12 @@
 use crate::system::{RadioClockControl, RadioClockController, RadioPeripherals};
 
-// Mask for clock bits used by both WIFI and Bluetooth, 0, 1, 2, 3, 7, 8, 9, 10,
-// 19, 20, 21, 22, 23
+// Note: this comment has been copied from esp-idf, including the mistake.
+// Mask for clock bits used by both WIFI and Bluetooth, 0, 1, 2, 3, 7,
+// 8, 9, 10, 19, 20, 21, 22, 23
 const SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M: u32 = 0x78078F;
+
 // SYSTEM_WIFI_CLK_EN : R/W ;bitpos:[31:0] ;default: 32'hfffce030
-// from experiments `0x00FB9FCF` is not enough for esp-wifi to work
-const SYSTEM_WIFI_CLK_EN: u32 = 0xFFFFFFFF;
+const SYSTEM_WIFI_CLK_EN: u32 = 0x00FB9FCF;
 
 impl RadioClockController for RadioClockControl {
     fn enable(&mut self, peripheral: RadioPeripherals) {
@@ -42,31 +43,27 @@ impl RadioClockController for RadioClockControl {
 }
 
 fn enable_phy() {
-    let system = unsafe { &*esp32s3::SYSTEM::PTR };
-    system
-        .perip_clk_en1
+    // `periph_ll_wifi_bt_module_enable_clk_clear_rst`
+    let syscon = unsafe { &*esp32s3::APB_CTRL::PTR };
+    syscon
+        .wifi_clk_en
         .modify(|r, w| unsafe { w.bits(r.bits() | SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M) });
 }
 
 fn disable_phy() {
-    let system = unsafe { &*esp32s3::SYSTEM::PTR };
-    system
-        .perip_clk_en1
+    // `periph_ll_wifi_bt_module_disable_clk_set_rst`
+    let syscon = unsafe { &*esp32s3::APB_CTRL::PTR };
+    syscon
+        .wifi_clk_en
         .modify(|r, w| unsafe { w.bits(r.bits() & !SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M) });
 }
 
 fn common_wifi_bt_clock_enable() {
-    let system = unsafe { &*esp32s3::SYSTEM::PTR };
-    system
-        .perip_clk_en1
-        .modify(|r, w| unsafe { w.bits(r.bits() | SYSTEM_WIFI_CLK_EN) });
+    // `periph_ll_wifi_module_enable_clk_clear_rst`. does nothing
 }
 
 fn common_wifi_bt_clock_disable() {
-    let system = unsafe { &*esp32s3::SYSTEM::PTR };
-    system
-        .perip_clk_en1
-        .modify(|r, w| unsafe { w.bits(r.bits() & !SYSTEM_WIFI_CLK_EN) });
+    // `periph_ll_wifi_module_disable_clk_set_rst`. does nothing
 }
 
 fn reset_mac() {
@@ -74,15 +71,23 @@ fn reset_mac() {
     let syscon = unsafe { &*esp32s3::APB_CTRL::PTR };
     syscon
         .wifi_rst_en
-        .modify(|r, w| unsafe { w.wifi_rst().bits(r.wifi_rst().bits() | SYSTEM_MAC_RST) });
+        .modify(|r, w| unsafe { w.bits(r.bits() | SYSTEM_MAC_RST) });
     syscon
         .wifi_rst_en
-        .modify(|r, w| unsafe { w.wifi_rst().bits(r.wifi_rst().bits() & !SYSTEM_MAC_RST) });
+        .modify(|r, w| unsafe { w.bits(r.bits() & !SYSTEM_MAC_RST) });
 }
 
 fn init_clocks() {
     let syscon = unsafe { &*esp32s3::APB_CTRL::PTR };
+
+    const SYSTEM_WIFI_CLK_I2C_CLK_EN: u32 = 1 << 5;
+    const SYSTEM_WIFI_CLK_UNUSED_BIT12: u32 = 1 << 12;
+    const SYSTEM_WIFI_CLK_SDIO_HOST_EN: u32 = 1 << 13;
+
+    const WIFI_BT_SDIO_CLK: u32 =
+        SYSTEM_WIFI_CLK_I2C_CLK_EN | SYSTEM_WIFI_CLK_UNUSED_BIT12 | SYSTEM_WIFI_CLK_SDIO_HOST_EN;
+
     syscon
         .wifi_clk_en
-        .modify(|_, w| unsafe { w.bits(0xffffffff) });
+        .modify(|r, w| unsafe { w.bits(r.bits() & !WIFI_BT_SDIO_CLK | SYSTEM_WIFI_CLK_EN) });
 }