+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add missing interrupt status read for esp32s3, which fixes USB-SERIAL-JTAG interrupts (#664)
 - GPIO interrupt status bits are now properly cleared (#670)
 - Increase frequency resolution in `set_periodic` (#686)
+- Fixed ESP32-S3 radio clock gating (#679)
 
 ### Removed